V0.4 (Major Update)
So we have added some new design features such asa tools div up top, so you can go and access basic tools whilst typing, those buttons show some preloaded fonts and sizes, the custom fonts and sizes however have not been touched, the menu now is functioning and i have moved, import .txt file, lock page and text to speech,  added the VividSpectrum menu (more down below).

VividSpectrum is a design language to do Googles MaterialYou one day. 

Update 0.2
We have built in a new menu with 6 new colours, Peach, Yellow, Magenta, Blue (Default), Green and red,  the design has now been finalised for the time being (as in the actual buttons and things not the colours) but the colours are still to come, and im working on something big with this.

If there are any issues, put it in the issues branch, also remember to put it in the same folder for everything.