V 0.6 Beta
We have added a highlight tool, rich text editing has been improved where you can fully add images, there is also an about button. Some placeholder setting buttons are now in the "VividSpectrum Design Menu" - Now just the themes tab (The VividSpectrum is not going away though). it also has been renamed to "settings" on the button.

This is a beta and things will change as time goes on, nothing is completely final. 

We are aiming for Dec for pre-release and late December for 1.0